+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Change Log
 
 ## [Unreleased]
+## [1.1.7] 2026-09-23
 
 ### Fixed